move pivot test data to a folder
