+++ test/example-types/metricGroups/pivotData/pivotRepsonseWithFilteredFieldValueGroup.js
