+++ test/example-types/metricGroups/pivotData/pivotResponse.js
